Backend/requirements.txt
Flask==2.1.1
Werkzeug==2.2.3
mysql-connector-python==8.0.28